# device: optimize Peer.String even more so