```
name          old time/op    new time/op    delta
PeerString-8    49.5ns ±17%    32.6ns ±17%  -34.09%  (p=0.000 n=10+10)

name          old alloc/op   new alloc/op   delta
PeerString-8     24.0B ± 0%     24.0B ± 0%     ~     (all equal)

name          old allocs/op  new allocs/op  delta
PeerString-8      1.00 ± 0%      1.00 ± 0%     ~     (all equal)
```

Signed-off-by: Alexander Yastrebov <yastrebov.alex@gmail.com>